feat: standardize CUDA code style and add vector_add vectorized

+++ ops/vector_add/include/vector_add.h
@@ -2,8 +2,4 @@
 
 const char* vector_add_variant_name();
 
-void launch_vector_add(const float* d_a,
-                       const float* d_b,
-                       float* d_out,
-                       int n);
-
+void launch_vector_add(const float* d_a, const float* d_b, float* d_out, size_t n);
